+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All breaking changes are marked with [BC] and potentially require API consumer changes after updating to the respective version.
 
+## Unreleased
+### Added
+- Add `Terminal::on_main_screen` for executing a function in a "normal" terminal state.
+
 ## [0.2.2] - 2019-10-24
 ### Added
 - Add `LineEdit::cursor_pos` to retrieve (byte) cursor position.